replace data preprocessing function wip
